web/requirements.txt
# Web Backend Dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0

# Existing project dependencies (if needed)
pyodbc>=4.0.34
python-dotenv>=1.0.0